refactor: move lazy import registry to _lazy_imports.py
Move _LAZY_IMPORT_REGISTRY and _get_lazy_import_registry() from
__init__.py to _lazy_imports.py to keep all lazy import logic
centralized in one module.

Benefits:
- Better code organization (all lazy import logic in one place)
- Cleaner __init__.py (reduced from ~80 lines to ~10 lines)
- Registry can directly reference name tuples and handlers in same module
- No need to import tuples/handlers when building registry
